test validateBody
